+++ secrets.go
@@ -6,6 +6,8 @@ package templig
 import (
 	"regexp"
 	"strings"
+
+	"gopkg.in/yaml.v3"
 )
 
 // SecretDefaultRE is the default regular expression used to identify secret values automatically.
@@ -15,47 +17,46 @@ const SecretDefaultRE = "(key)|(secret)|(pass)|(password)|(cert)|(certificate)"
 // In case there are different properties to identify secrets, extend it.
 var SecretRE = regexp.MustCompile(SecretDefaultRE)
 
-// hideSecretsStringMap hides secrets inside a map of string keys. It is the only place secrets can
-// be detected due to their naming inside the map.
-func hideSecretsStringMap(m map[string]any) {
-	for k, v := range m {
-		if SecretRE.MatchString(strings.ToLower(k)) {
-			switch vt := v.(type) {
-			case string:
-				m[k] = strings.Repeat("*", len(vt))
-			default:
-				m[k] = "*"
+// HideSecrets hides secrets in the given YAML node structure. Secrets are identified using the [SecretRE].
+// Depending on the parameter `hideStructure` the structure of the secret is hidden too (`true`) or visible (`false`).
+func HideSecrets(n *yaml.Node, hideStructure bool) {
+	if n == nil {
+		return
+	}
+
+	if n.Kind == yaml.MappingNode {
+		for i := 0; i < len(n.Content); i += 2 {
+			if SecretRE.Match([]byte(n.Content[i].Value)) {
+				hideAll(n.Content[i+1], hideStructure)
+			} else {
+				HideSecrets(n.Content[i+1], hideStructure)
 			}
-		} else {
-			hideSecrets(v)
+		}
+	} else {
+		for _, v := range n.Content {
+			HideSecrets(v, hideStructure)
 		}
 	}
 }
 
-// hideSecretsAnyMap hides secrets inside a non-string keyed map. That means, that secrets can
-// only be inside the substructure, or they would be hidden by the [hideSecretsStringMap] function.
-func hideSecretsAnyMap(m map[any]any) {
-	for _, v := range m {
-		hideSecrets(v)
-	}
-}
-
-// hideSecrets hides secrets inside the value given. It analyses the basic structure and decides which
-// one of the specialized hiding functions to apply.
-func hideSecrets(a any) {
-	switch ta := a.(type) {
-	case map[string]any:
-		hideSecretsStringMap(ta)
-	case map[any]any:
-		hideSecretsAnyMap(ta)
-	case []any:
-		hideSecretsSlice(ta)
-	}
-}
-
-// hideSecretsSlice hides secrets inside the given slice. It applies the [hideSecrets] function on each value.
-func hideSecretsSlice(s []any) {
-	for _, v := range s {
-		hideSecrets(v)
+func hideAll(n *yaml.Node, hideStructure bool) {
+	if n.Kind == yaml.ScalarNode {
+		n.Tag = "!!str"
+		n.Value = strings.Repeat("*", len(n.Value))
+	} else if n.Kind == yaml.AliasNode {
+		if n.Alias != nil {
+			hideAll(n.Alias, hideStructure)
+		}
+	} else {
+		if hideStructure {
+			n.Kind = yaml.ScalarNode
+			n.Tag = "!!str"
+			n.Value = "*"
+			n.Content = nil
+		} else {
+			for _, v := range n.Content {
+				hideAll(v, hideStructure)
+			}
+		}
 	}
 }